refactor!: remove dead public graphics surface (ICanvas, INode, GraphicsException, FpsText/DrawFps)
ICanvas/INode interfaces have a single implementer each and are never used polymorphically; GraphicsException is never thrown; the FPS overlay is permanently disabled (DrawFps is never set). Remove them, keeping the live types co-located in IBackdrop.cs.

BREAKING CHANGE: removes public Beutl.Graphics.ICanvas, Beutl.Graphics.Rendering.INode, Beutl.Graphics.GraphicsException, and IRenderer.DrawFps from Beutl.Engine.
