--- src/Beutl.Engine/Graphics/Rendering/INode.cs
@@ -1,6 +1,0 @@
-﻿namespace Beutl.Graphics.Rendering;
-
-public interface INode : IDisposable
-{
-    bool IsDisposed { get; }
-}
